fts: extract reader/metadata.rs (NormTable, ColumnMeta, OpenOptions)
Split the index-metadata + open-config types — the per-doc NormTable,
per-column ColumnMeta and its FtsColumnConfig JSON, default_tokenizer,
and the reader OpenOptions — out of core into reader/metadata.rs.
OpenOptions re-exported from mod.rs (fts::reader::OpenOptions is used
across the crate); NormTable imported directly by phrase/sink/scorers;
NormTable::new and default_tokenizer widened to pub(super). Pure move.
